pyproject.toml
[project]
name = "kaboom"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "influxdb-client>=1.49.0",
    "influxdb3-python>=0.16.0",
    "pandas>=2.3.2",
    "python-dotenv>=1.1.1",
    "supabase>=2.19.0",
    "toml>=0.10.2",
    "yfinance>=0.2.66",
]
